backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
cassandra-driver==3.28.0
neo4j==5.14.0
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pyspark==3.5.0
APScheduler==3.10.4
httpx==0.25.2
faker==20.1.0

